feat: Add missing person data for Martin Elgarno
This commit adds data for a new missing person, Martin Elgarno, to the `missing.json` file. The data includes details such as his ID, name, date of birth, assignment, rank, status, date of disappearance, last contact, resume, and annexes. This information will be used to populate the missing persons section of the application.